+++ webapp/src/components/ListPage/index.ts
@@ -1,0 +1,2 @@
+import ListPage from './ListPage.container'
+export { ListPage }
