--- webapp/src/components/ListsPage/index.ts
@@ -1,2 +1,0 @@
-import ListsPage from './ListsPage.container'
-export { ListsPage }
