+++ webapp/src/components/Routes/Routes.tsx
@@ -26,7 +26,7 @@ import { ManageAssetPage } from '../ManageAssetPage'
 import { Footer } from '../Footer'
 import { CampaignBrowserPage } from '../Campaign/CampaignBrowserPage'
 import { StatusPage } from '../BuyPage/StatusPage'
-import { ListsPage } from '../ListsPage'
+import { ListPage } from '../ListPage'
 import { Props } from './Routes.types'
 
 const Routes = ({ inMaintenance }: Props) => {
@@ -56,7 +56,7 @@ const Routes = ({ inMaintenance }: Props) => {
           component={AccountPage}
         />
         <Route exact path={locations.account()} component={AccountPage} />
-        <Route exact path={locations.list()} component={ListsPage} />
+        <Route exact path={locations.list()} component={ListPage} />
         <Route exact path={locations.signIn()} component={SignInPage} />
         <Route exact path={locations.sell()} component={SellPage} />
         <Route exact path={locations.bid()} component={BidPage} />
